+++ src/bandcamp/pages/page-music.js
@@ -5,8 +5,10 @@ import {
   createElementFromHTML,
   input,
   listenForMessage,
+  onClick,
   selectElementWithContent,
-  setDataAttribute
+  setDataAttribute,
+  triggerInputEvent
 } from '../../utils/html';
 import {
   containsOneOf,
@@ -16,6 +18,7 @@ import {
   removeBrackets
 } from '../../utils/utils';
 import {
+  extractBCSearchInputStyle,
   getBandPhotoSrc,
   getReleaseItems as getReleaseItemsFromPage
 } from '../modules/html.js';
@@ -109,12 +112,13 @@ function setupIsotope() {
     );
   });
 
-  const artistFilterWidget = createArtistFilterWidget(releaseItems);
+  const bcStyle = extractBCSearchInputStyle();
+  const artistFilterWidget = createArtistFilterWidget(releaseItems, bcStyle);
+  const albumAmountWidget = createAlbumAmountWidget(releaseItems, bcStyle);
+
   const filterBlock = createElementFromHTML(
     `<div class="b2d-widget-container"></div>`
   );
-  const albumAmountWidget = createAlbumAmountWidget(releaseItems);
-
   filterBlock.append(artistFilterWidget);
   filterBlock.append(albumAmountWidget);
 
@@ -227,7 +231,7 @@ function getArtistListData(releaseItems) {
   return [...new Set(filterData)];
 }
 
-function createArtistFilterWidget(releaseItems) {
+function createArtistFilterWidget(releaseItems, bcStyle) {
   let artistFilterElement = createElementFromHTML(
     `<div class="b2d-widget">
   <label for="b2dArtistFilter">Artist / Album:</label>
@@ -241,24 +245,41 @@ function createArtistFilterWidget(releaseItems) {
   );
   const artistFilterInput =
     artistFilterElement.querySelector('#b2dArtistFilter');
-
-  // Extract background-color from the BC input element
-  const bcSearchInputElement = document.querySelector('input.search-bar');
-  const bcSearchInputElementStyle =
-    window.getComputedStyle(bcSearchInputElement);
-  artistFilterInput.style.backgroundColor =
-    bcSearchInputElementStyle.backgroundColor;
-  artistFilterInput.style.color = bcSearchInputElementStyle.color;
+  artistFilterInput.style.backgroundColor = bcStyle.backgroundColor;
+  artistFilterInput.style.color = bcStyle.color;
 
   artistFilterElement.append(artistFilterDatalist);
+
+  const clearButton = createElementFromHTML(
+    `<button id="b2dArtistFilterClear" title="Clear the filter">
+<svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-x-circle" viewBox="0 0 16 16">
+  <path d="M8 15A7 7 0 1 1 8 1a7 7 0 0 1 0 14m0 1A8 8 0 1 0 8 0a8 8 0 0 0 0 16"/>
+  <path d="M4.646 4.646a.5.5 0 0 1 .708 0L8 7.293l2.646-2.647a.5.5 0 0 1 .708.708L8.707 8l2.647 2.646a.5.5 0 0 1-.708.708L8 8.707l-2.646 2.647a.5.5 0 0 1-.708-.708L7.293 8 4.646 5.354a.5.5 0 0 1 0-.708"/>
+</svg>
+  </button>`
+  );
+
+  onClick(clearButton, () => {
+    artistFilterInput.value = '';
+    triggerInputEvent(artistFilterInput);
+  });
+
+  artistFilterElement.append(clearButton);
 
   return artistFilterElement;
 }
 
-function createAlbumAmountWidget(releaseItems) {
-  return createElementFromHTML(
+function createAlbumAmountWidget(releaseItems, bcStyle) {
+  const widget = createElementFromHTML(
     `<div class="b2d-albumAmount b2d-widget" title="The displayed and total amount of albums on the page">
 Displayed: <span class="b2d-badge b2d-visible">${releaseItems.length}</span> Total: <span class="b2d-badge b2d-total">${releaseItems.length}</span>
 </div>`
   );
+  const badges = widget.querySelectorAll('.b2d-badge');
+  badges?.forEach((badge) => {
+    badge.style.backgroundColor = bcStyle.backgroundColor;
+    badge.style.color = bcStyle.color;
+  });
+
+  return widget;
 }